@@ -2,8 +2,6 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-// @dart = 2.8
-
 import 'dart:ui' as ui show Image;
 
 import 'box.dart';
@@ -27,19 +25,19 @@ class RenderImage extends RenderBox {
   /// must not be null. The [textDirection] argument must not be null if
   /// [alignment] will need resolving or if [matchTextDirection] is true.
   RenderImage({
-    ui.Image image,
+    ui.Image? image,
     this.debugImageLabel,
-    double width,
-    double height,
+    double? width,
+    double? height,
     double scale = 1.0,
-    Color color,
-    BlendMode colorBlendMode,
-    BoxFit fit,
+    Color? color,
+    BlendMode? colorBlendMode,
+    BoxFit? fit,
     AlignmentGeometry alignment = Alignment.center,
     ImageRepeat repeat = ImageRepeat.noRepeat,
-    Rect centerSlice,
+    Rect? centerSlice,
     bool matchTextDirection = false,
-    TextDirection textDirection,
+    TextDirection? textDirection,
     bool invertColors = false,
     bool isAntiAlias = false,
     FilterQuality filterQuality = FilterQuality.low,
@@ -67,8 +65,8 @@ class RenderImage extends RenderBox {
     _updateColorFilter();
   }
 
-  Alignment _resolvedAlignment;
-  bool _flipHorizontally;
+  Alignment? _resolvedAlignment;
+  bool? _flipHorizontally;
 
   void _resolve() {
     if (_resolvedAlignment != null)
@@ -84,9 +82,9 @@ class RenderImage extends RenderBox {
   }
 
   /// The image to display.
-  ui.Image get image => _image;
-  ui.Image _image;
-  set image(ui.Image value) {
+  ui.Image? get image => _image;
+  ui.Image? _image;
+  set image(ui.Image? value) {
     if (value == _image)
       return;
     _image = value;
@@ -96,15 +94,15 @@ class RenderImage extends RenderBox {
   }
 
   /// A string used to identify the source of the image.
-  String debugImageLabel;
+  String? debugImageLabel;
 
   /// If non-null, requires the image to have this width.
   ///
   /// If null, the image will pick a size that best preserves its intrinsic
   /// aspect ratio.
-  double get width => _width;
-  double _width;
-  set width(double value) {
+  double? get width => _width;
+  double? _width;
+  set width(double? value) {
     if (value == _width)
       return;
     _width = value;
@@ -115,9 +113,9 @@ class RenderImage extends RenderBox {
   ///
   /// If null, the image will pick a size that best preserves its intrinsic
   /// aspect ratio.
-  double get height => _height;
-  double _height;
-  set height(double value) {
+  double? get height => _height;
+  double? _height;
+  set height(double? value) {
     if (value == _height)
       return;
     _height = value;
@@ -137,19 +135,19 @@ class RenderImage extends RenderBox {
     markNeedsLayout();
   }
 
-  ColorFilter _colorFilter;
+  ColorFilter? _colorFilter;
 
   void _updateColorFilter() {
     if (_color == null)
       _colorFilter = null;
     else
-      _colorFilter = ColorFilter.mode(_color, _colorBlendMode ?? BlendMode.srcIn);
+      _colorFilter = ColorFilter.mode(_color!, _colorBlendMode ?? BlendMode.srcIn);
   }
 
   /// If non-null, this color is blended with each image pixel using [colorBlendMode].
-  Color get color => _color;
-  Color _color;
-  set color(Color value) {
+  Color? get color => _color;
+  Color? _color;
+  set color(Color? value) {
     if (value == _color)
       return;
     _color = value;
@@ -180,9 +178,9 @@ class RenderImage extends RenderBox {
   /// See also:
   ///
   ///  * [BlendMode], which includes an illustration of the effect of each blend mode.
-  BlendMode get colorBlendMode => _colorBlendMode;
-  BlendMode _colorBlendMode;
-  set colorBlendMode(BlendMode value) {
+  BlendMode? get colorBlendMode => _colorBlendMode;
+  BlendMode? _colorBlendMode;
+  set colorBlendMode(BlendMode? value) {
     if (value == _colorBlendMode)
       return;
     _colorBlendMode = value;
@@ -194,9 +192,9 @@ class RenderImage extends RenderBox {
   ///
   /// The default varies based on the other fields. See the discussion at
   /// [paintImage].
-  BoxFit get fit => _fit;
-  BoxFit _fit;
-  set fit(BoxFit value) {
+  BoxFit? get fit => _fit;
+  BoxFit? _fit;
+  set fit(BoxFit? value) {
     if (value == _fit)
       return;
     _fit = value;
@@ -235,9 +233,9 @@ class RenderImage extends RenderBox {
   /// region of the image above and below the center slice will be stretched
   /// only horizontally and the region of the image to the left and right of
   /// the center slice will be stretched only vertically.
-  Rect get centerSlice => _centerSlice;
-  Rect _centerSlice;
-  set centerSlice(Rect value) {
+  Rect? get centerSlice => _centerSlice;
+  Rect? _centerSlice;
+  set centerSlice(Rect? value) {
     if (value == _centerSlice)
       return;
     _centerSlice = value;
@@ -287,9 +285,9 @@ class RenderImage extends RenderBox {
   /// This may be changed to null, but only after the [alignment] and
   /// [matchTextDirection] properties have been changed to values that do not
   /// depend on the direction.
-  TextDirection get textDirection => _textDirection;
-  TextDirection _textDirection;
-  set textDirection(TextDirection value) {
+  TextDirection? get textDirection => _textDirection;
+  TextDirection? _textDirection;
+  set textDirection(TextDirection? value) {
     if (_textDirection == value)
       return;
     _textDirection = value;
@@ -329,8 +327,8 @@ class RenderImage extends RenderBox {
       return constraints.smallest;
 
     return constraints.constrainSizeAndAttemptToPreserveAspectRatio(Size(
-      _image.width.toDouble() / _scale,
-      _image.height.toDouble() / _scale,
+      _image!.width.toDouble() / _scale,
+      _image!.height.toDouble() / _scale,
     ));
   }
 
@@ -380,15 +378,15 @@ class RenderImage extends RenderBox {
     paintImage(
       canvas: context.canvas,
       rect: offset & size,
-      image: _image,
+      image: _image!,
       debugImageLabel: debugImageLabel,
       scale: _scale,
       colorFilter: _colorFilter,
       fit: _fit,
-      alignment: _resolvedAlignment,
+      alignment: _resolvedAlignment!,
       centerSlice: _centerSlice,
       repeat: _repeat,
-      flipHorizontally: _flipHorizontally,
+      flipHorizontally: _flipHorizontally!,
       invertColors: invertColors,
       filterQuality: _filterQuality,
       isAntiAlias: _isAntiAlias,
