@@ -2,11 +2,8 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-// @dart = 2.8
-
 import 'dart:math' as math;
 
-import 'package:flutter/foundation.dart';
 import 'package:flutter/gestures.dart';
 import 'package:flutter/painting.dart';
 import 'package:vector_math/vector_math_64.dart';
@@ -26,8 +23,8 @@ class RenderRotatedBox extends RenderBox with RenderObjectWithChildMixin<RenderB
   ///
   /// The [quarterTurns] argument must not be null.
   RenderRotatedBox({
-    @required int quarterTurns,
-    RenderBox child,
+    required int quarterTurns,
+    RenderBox? child,
   }) : assert(quarterTurns != null),
        _quarterTurns = quarterTurns {
     this.child = child;
@@ -50,75 +47,75 @@ class RenderRotatedBox extends RenderBox with RenderObjectWithChildMixin<RenderB
   double computeMinIntrinsicWidth(double height) {
     if (child == null)
       return 0.0;
-    return _isVertical ? child.getMinIntrinsicHeight(height) : child.getMinIntrinsicWidth(height);
+    return _isVertical ? child!.getMinIntrinsicHeight(height) : child!.getMinIntrinsicWidth(height);
   }
 
   @override
   double computeMaxIntrinsicWidth(double height) {
     if (child == null)
       return 0.0;
-    return _isVertical ? child.getMaxIntrinsicHeight(height) : child.getMaxIntrinsicWidth(height);
+    return _isVertical ? child!.getMaxIntrinsicHeight(height) : child!.getMaxIntrinsicWidth(height);
   }
 
   @override
   double computeMinIntrinsicHeight(double width) {
     if (child == null)
       return 0.0;
-    return _isVertical ? child.getMinIntrinsicWidth(width) : child.getMinIntrinsicHeight(width);
+    return _isVertical ? child!.getMinIntrinsicWidth(width) : child!.getMinIntrinsicHeight(width);
   }
 
   @override
   double computeMaxIntrinsicHeight(double width) {
     if (child == null)
       return 0.0;
-    return _isVertical ? child.getMaxIntrinsicWidth(width) : child.getMaxIntrinsicHeight(width);
+    return _isVertical ? child!.getMaxIntrinsicWidth(width) : child!.getMaxIntrinsicHeight(width);
   }
 
-  Matrix4 _paintTransform;
+  Matrix4? _paintTransform;
 
   @override
   void performLayout() {
     _paintTransform = null;
     if (child != null) {
-      child.layout(_isVertical ? constraints.flipped : constraints, parentUsesSize: true);
-      size = _isVertical ? Size(child.size.height, child.size.width) : child.size;
+      child!.layout(_isVertical ? constraints.flipped : constraints, parentUsesSize: true);
+      size = _isVertical ? Size(child!.size.height, child!.size.width) : child!.size;
       _paintTransform = Matrix4.identity()
         ..translate(size.width / 2.0, size.height / 2.0)
         ..rotateZ(_kQuarterTurnsInRadians * (quarterTurns % 4))
-        ..translate(-child.size.width / 2.0, -child.size.height / 2.0);
+        ..translate(-child!.size.width / 2.0, -child!.size.height / 2.0);
     } else {
       performResize();
     }
   }
 
   @override
-  bool hitTestChildren(BoxHitTestResult result, { Offset position }) {
+  bool hitTestChildren(BoxHitTestResult result, { required Offset position }) {
     assert(_paintTransform != null || debugNeedsLayout || child == null);
     if (child == null || _paintTransform == null)
       return false;
     return result.addWithPaintTransform(
       transform: _paintTransform,
       position: position,
-      hitTest: (BoxHitTestResult result, Offset position) {
-        return child.hitTest(result, position: position);
+      hitTest: (BoxHitTestResult result, Offset? position) {
+        return child!.hitTest(result, position: position!);
       },
     );
   }
 
   void _paintChild(PaintingContext context, Offset offset) {
-    context.paintChild(child, offset);
+    context.paintChild(child!, offset);
   }
 
   @override
   void paint(PaintingContext context, Offset offset) {
     if (child != null)
-      context.pushTransform(needsCompositing, offset, _paintTransform, _paintChild);
+      context.pushTransform(needsCompositing, offset, _paintTransform!, _paintChild);
   }
 
   @override
   void applyPaintTransform(RenderBox child, Matrix4 transform) {
     if (_paintTransform != null)
-      transform.multiply(_paintTransform);
+      transform.multiply(_paintTransform!);
     super.applyPaintTransform(child, transform);
   }
 }
