@@ -2,8 +2,6 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-// @dart = 2.8
-
 import 'dart:collection';
 import 'dart:math' as math;
 
@@ -16,13 +14,13 @@ import 'table_border.dart';
 /// Parent data used by [RenderTable] for its children.
 class TableCellParentData extends BoxParentData {
   /// Where this cell should be placed vertically.
-  TableCellVerticalAlignment verticalAlignment;
+  TableCellVerticalAlignment? verticalAlignment;
 
   /// The column that the child was in the last time it was laid out.
-  int x;
+  int? x;
 
   /// The row that the child was in the last time it was laid out.
-  int y;
+  int? y;
 
   @override
   String toString() => '${super.toString()}; ${verticalAlignment == null ? "default vertical alignment" : "$verticalAlignment"}';
@@ -75,7 +73,7 @@ abstract class TableColumnWidth {
   /// The `cells` argument is an iterable that provides all the cells
   /// in the table for this column. Walking the cells is by definition
   /// O(N), so algorithms that do that should be considered expensive.
-  double flex(Iterable<RenderBox> cells) => null;
+  double? flex(Iterable<RenderBox> cells) => null;
 
   @override
   String toString() => objectRuntimeType(this, 'TableColumnWidth');
@@ -98,7 +96,7 @@ class IntrinsicColumnWidth extends TableColumnWidth {
   /// there is any room left over when laying out the table. If `flex` is
   /// null (the default), the table will not distribute any extra space to the
   /// column.
-  const IntrinsicColumnWidth({ double flex }) : _flex = flex;
+  const IntrinsicColumnWidth({ double? flex }) : _flex = flex;
 
   @override
   double minIntrinsicWidth(Iterable<RenderBox> cells, double containerWidth) {
@@ -116,10 +114,10 @@ class IntrinsicColumnWidth extends TableColumnWidth {
     return result;
   }
 
-  final double _flex;
+  final double? _flex;
 
   @override
-  double flex(Iterable<RenderBox> cells) => _flex;
+  double? flex(Iterable<RenderBox> cells) => _flex;
 
   @override
   String toString() => '${objectRuntimeType(this, 'IntrinsicColumnWidth')}(flex: ${_flex?.toStringAsFixed(1)})';
@@ -257,11 +255,11 @@ class MaxColumnWidth extends TableColumnWidth {
   }
 
   @override
-  double flex(Iterable<RenderBox> cells) {
-    final double aFlex = a.flex(cells);
+  double? flex(Iterable<RenderBox> cells) {
+    final double? aFlex = a.flex(cells);
     if (aFlex == null)
       return b.flex(cells);
-    final double bFlex = b.flex(cells);
+    final double? bFlex = b.flex(cells);
     if (bFlex == null)
       return null;
     return math.max(aFlex, bFlex);
@@ -308,11 +306,11 @@ class MinColumnWidth extends TableColumnWidth {
   }
 
   @override
-  double flex(Iterable<RenderBox> cells) {
-    final double aFlex = a.flex(cells);
+  double? flex(Iterable<RenderBox> cells) {
+    final double? aFlex = a.flex(cells);
     if (aFlex == null)
       return b.flex(cells);
-    final double bFlex = b.flex(cells);
+    final double? bFlex = b.flex(cells);
     if (bFlex == null)
       return null;
     return math.min(aFlex, bFlex);
@@ -362,43 +360,44 @@ class RenderTable extends RenderBox {
   ///    null, then `children` must be null.
   ///  * `children` must either be null or contain lists of all the same length.
   ///    if `children` is not null, then `rows` must be null.
+  ///  * [columnWidths] may be null, in which case it defaults to an empty map.
   ///  * [defaultColumnWidth] must not be null.
   ///  * [configuration] must not be null (but has a default value).
   RenderTable({
-    int columns,
-    int rows,
-    Map<int, TableColumnWidth> columnWidths,
+    int? columns,
+    int? rows,
+    Map<int, TableColumnWidth>? columnWidths,
     TableColumnWidth defaultColumnWidth = const FlexColumnWidth(1.0),
-    @required TextDirection textDirection,
-    TableBorder border,
-    List<Decoration> rowDecorations,
+    required TextDirection textDirection,
+    TableBorder? border,
+    List<Decoration?>? rowDecorations,
     ImageConfiguration configuration = ImageConfiguration.empty,
     TableCellVerticalAlignment defaultVerticalAlignment = TableCellVerticalAlignment.top,
-    TextBaseline textBaseline,
-    List<List<RenderBox>> children,
+    TextBaseline? textBaseline,
+    List<List<RenderBox>>? children,
   }) : assert(columns == null || columns >= 0),
        assert(rows == null || rows >= 0),
        assert(rows == null || children == null),
        assert(defaultColumnWidth != null),
        assert(textDirection != null),
        assert(configuration != null),
-       _textDirection = textDirection {
-    _columns = columns ?? (children != null && children.isNotEmpty ? children.first.length : 0);
-    _rows = rows ?? 0;
+       _textDirection = textDirection,
+       _columns = columns ?? (children != null && children.isNotEmpty ? children.first.length : 0),
+       _rows = rows ?? 0,
+       _columnWidths = columnWidths ?? HashMap<int, TableColumnWidth>(),
+       _defaultColumnWidth = defaultColumnWidth,
+       _border = border,
+       _textBaseline = textBaseline,
+       _defaultVerticalAlignment = defaultVerticalAlignment,
+       _configuration = configuration {
     _children = <RenderBox>[]..length = _columns * _rows;
-    _columnWidths = columnWidths ?? HashMap<int, TableColumnWidth>();
-    _defaultColumnWidth = defaultColumnWidth;
-    _border = border;
     this.rowDecorations = rowDecorations; // must use setter to initialize box painters array
-    _configuration = configuration;
-    _defaultVerticalAlignment = defaultVerticalAlignment;
-    _textBaseline = textBaseline;
     children?.forEach(addRow);
   }
 
   // Children are stored in row-major order.
   // _children.length must be rows * columns
-  List<RenderBox> _children = const <RenderBox>[];
+  List<RenderBox?> _children = const <RenderBox?>[];
 
   /// The number of vertical alignment lines in this table.
   ///
@@ -415,9 +414,9 @@ class RenderTable extends RenderBox {
     if (value == columns)
       return;
     final int oldColumns = columns;
-    final List<RenderBox> oldChildren = _children;
+    final List<RenderBox?> oldChildren = _children;
     _columns = value;
-    _children = <RenderBox>[]..length = columns * rows;
+    _children = List<RenderBox?>.filled(columns * rows, null, growable: false);
     final int columnsToCopy = math.min(columns, oldColumns);
     for (int y = 0; y < rows; y += 1) {
       for (int x = 0; x < columnsToCopy; x += 1)
@@ -428,7 +427,7 @@ class RenderTable extends RenderBox {
         for (int x = columns; x < oldColumns; x += 1) {
           final int xy = x + y * oldColumns;
           if (oldChildren[xy] != null)
-            dropChild(oldChildren[xy]);
+            dropChild(oldChildren[xy]!);
         }
       }
     }
@@ -449,7 +448,7 @@ class RenderTable extends RenderBox {
     if (_rows > value) {
       for (int xy = columns * value; xy < _children.length; xy += 1) {
         if (_children[xy] != null)
-          dropChild(_children[xy]);
+          dropChild(_children[xy]!);
       }
     }
     _rows = value;
@@ -466,13 +465,18 @@ class RenderTable extends RenderBox {
   /// sizing algorithms are used here. In particular, [IntrinsicColumnWidth] is
   /// quite expensive because it needs to measure each cell in the column to
   /// determine the intrinsic size of the column.
-  Map<int, TableColumnWidth> get columnWidths => Map<int, TableColumnWidth>.unmodifiable(_columnWidths);
+  ///
+  /// This property can never return null. If it is set to null, and the existing
+  /// map is not empty, then the value is replaced by an empty map. (If it is set
+  /// to null while the current value is an empty map, the value is not changed.)
+  Map<int, TableColumnWidth>? get columnWidths => Map<int, TableColumnWidth>.unmodifiable(_columnWidths);
   Map<int, TableColumnWidth> _columnWidths;
-  set columnWidths(Map<int, TableColumnWidth> value) {
-    value ??= HashMap<int, TableColumnWidth>();
+  set columnWidths(Map<int, TableColumnWidth>? value) {
     if (_columnWidths == value)
       return;
-    _columnWidths = value;
+    if (_columnWidths.isEmpty && value == null)
+      return;
+    _columnWidths = value ?? HashMap<int, TableColumnWidth>();
     markNeedsLayout();
   }
 
@@ -510,9 +514,9 @@ class RenderTable extends RenderBox {
   }
 
   /// The style to use when painting the boundary and interior divisions of the table.
-  TableBorder get border => _border;
-  TableBorder _border;
-  set border(TableBorder value) {
+  TableBorder? get border => _border;
+  TableBorder? _border;
+  set border(TableBorder? value) {
     if (border == value)
       return;
     _border = value;
@@ -527,17 +531,17 @@ class RenderTable extends RenderBox {
   List<Decoration> get rowDecorations => List<Decoration>.unmodifiable(_rowDecorations ?? const <Decoration>[]);
   // _rowDecorations and _rowDecorationPainters need to be in sync. They have to
   // either both be null or have same length.
-  List<Decoration> _rowDecorations;
-  List<BoxPainter> _rowDecorationPainters;
-  set rowDecorations(List<Decoration> value) {
+  List<Decoration?>? _rowDecorations;
+  List<BoxPainter?>? _rowDecorationPainters;
+  set rowDecorations(List<Decoration?>? value) {
     if (_rowDecorations == value)
       return;
     _rowDecorations = value;
     if (_rowDecorationPainters != null) {
-      for (final BoxPainter painter in _rowDecorationPainters)
+      for (final BoxPainter? painter in _rowDecorationPainters!)
         painter?.dispose();
     }
-    _rowDecorationPainters = _rowDecorations != null ? List<BoxPainter>(_rowDecorations.length) : null;
+    _rowDecorationPainters = _rowDecorations != null ? List<BoxPainter?>.filled(_rowDecorations!.length, null, growable: false) : null;
   }
 
   /// The settings to pass to the [rowDecorations] when painting, so that they
@@ -557,6 +561,7 @@ class RenderTable extends RenderBox {
   TableCellVerticalAlignment get defaultVerticalAlignment => _defaultVerticalAlignment;
   TableCellVerticalAlignment _defaultVerticalAlignment;
   set defaultVerticalAlignment(TableCellVerticalAlignment value) {
+    assert(value != null);
     if (_defaultVerticalAlignment == value)
       return;
     _defaultVerticalAlignment = value;
@@ -564,9 +569,9 @@ class RenderTable extends RenderBox {
   }
 
   /// The text baseline to use when aligning rows using [TableCellVerticalAlignment.baseline].
-  TextBaseline get textBaseline => _textBaseline;
-  TextBaseline _textBaseline;
-  set textBaseline(TextBaseline value) {
+  TextBaseline? get textBaseline => _textBaseline;
+  TextBaseline? _textBaseline;
+  set textBaseline(TextBaseline? value) {
     if (_textBaseline == value)
       return;
     _textBaseline = value;
@@ -587,7 +592,7 @@ class RenderTable extends RenderBox {
   /// If the new cells contain any existing children of the table, those
   /// children are simply moved to their new location in the table rather than
   /// removed from the table and re-added.
-  void setFlatChildren(int columns, List<RenderBox> cells) {
+  void setFlatChildren(int columns, List<RenderBox?> cells) {
     if (cells == _children && columns == _columns)
       return;
     assert(columns >= 0);
@@ -599,7 +604,7 @@ class RenderTable extends RenderBox {
         assert(_rows == 0);
         return;
       }
-      for (final RenderBox oldChild in _children) {
+      for (final RenderBox? oldChild in _children) {
         if (oldChild != null)
           dropChild(oldChild);
       }
@@ -619,7 +624,7 @@ class RenderTable extends RenderBox {
         final int xyOld = x + y * _columns;
         final int xyNew = x + y * columns;
         if (_children[xyOld] != null && (x >= columns || xyNew >= cells.length || _children[xyOld] != cells[xyNew]))
-          lostChildren.add(_children[xyOld]);
+          lostChildren.add(_children[xyOld]!);
       }
     }
     // adopt cells that are arriving, and cross cells that are just moving off our list of lostChildren
@@ -630,7 +635,7 @@ class RenderTable extends RenderBox {
         final int xyOld = x + y * _columns;
         if (cells[xyNew] != null && (x >= _columns || y >= _rows || _children[xyOld] != cells[xyNew])) {
           if (!lostChildren.remove(cells[xyNew]))
-            adoptChild(cells[xyNew]);
+            adoptChild(cells[xyNew]!);
         }
       }
       y += 1;
@@ -646,13 +651,13 @@ class RenderTable extends RenderBox {
   }
 
   /// Replaces the children of this table with the given cells.
-  void setChildren(List<List<RenderBox>> cells) {
+  void setChildren(List<List<RenderBox>>? cells) {
     // TODO(ianh): Make this smarter, like setFlatChildren
     if (cells == null) {
-      setFlatChildren(0, null);
+      setFlatChildren(0, const <RenderBox?>[]);
       return;
     }
-    for (final RenderBox oldChild in _children) {
+    for (final RenderBox? oldChild in _children) {
       if (oldChild != null)
         dropChild(oldChild);
     }
@@ -666,12 +671,12 @@ class RenderTable extends RenderBox {
   /// Adds a row to the end of the table.
   ///
   /// The newly added children must not already have parents.
-  void addRow(List<RenderBox> cells) {
+  void addRow(List<RenderBox?> cells) {
     assert(cells.length == columns);
     assert(_children.length == rows * columns);
     _rows += 1;
     _children.addAll(cells);
-    for (final RenderBox cell in cells) {
+    for (final RenderBox? cell in cells) {
       if (cell != null)
         adoptChild(cell);
     }
@@ -683,13 +688,13 @@ class RenderTable extends RenderBox {
   /// If the given child is already located at the given position, this function
   /// does not modify the table. Otherwise, the given child must not already
   /// have a parent.
-  void setChild(int x, int y, RenderBox value) {
+  void setChild(int x, int y, RenderBox? value) {
     assert(x != null);
     assert(y != null);
     assert(x >= 0 && x < columns && y >= 0 && y < rows);
     assert(_children.length == rows * columns);
     final int xy = x + y * columns;
-    final RenderBox oldChild = _children[xy];
+    final RenderBox? oldChild = _children[xy];
     if (oldChild == value)
       return;
     if (oldChild != null)
@@ -702,7 +707,7 @@ class RenderTable extends RenderBox {
   @override
   void attach(PipelineOwner owner) {
     super.attach(owner);
-    for (final RenderBox child in _children)
+    for (final RenderBox? child in _children)
       child?.attach(owner);
   }
 
@@ -710,18 +715,18 @@ class RenderTable extends RenderBox {
   void detach() {
     super.detach();
     if (_rowDecorationPainters != null) {
-      for (final BoxPainter painter in _rowDecorationPainters)
+      for (final BoxPainter? painter in _rowDecorationPainters!)
         painter?.dispose();
-      _rowDecorationPainters = List<BoxPainter>(_rowDecorations.length);
+      _rowDecorationPainters = List<BoxPainter?>.filled(_rowDecorations!.length, null, growable: false);
     }
-    for (final RenderBox child in _children)
+    for (final RenderBox? child in _children)
       child?.detach();
   }
 
   @override
   void visitChildren(RenderObjectVisitor visitor) {
     assert(_children.length == rows * columns);
-    for (final RenderBox child in _children) {
+    for (final RenderBox? child in _children) {
       if (child != null)
         visitor(child);
     }
@@ -762,7 +767,7 @@ class RenderTable extends RenderBox {
       double rowHeight = 0.0;
       for (int x = 0; x < columns; x += 1) {
         final int xy = x + y * columns;
-        final RenderBox child = _children[xy];
+        final RenderBox? child = _children[xy];
         if (child != null)
           rowHeight = math.max(rowHeight, child.getMaxIntrinsicHeight(widths[x]));
       }
@@ -776,9 +781,9 @@ class RenderTable extends RenderBox {
     return computeMinIntrinsicHeight(width);
   }
 
-  double _baselineDistance;
+  double? _baselineDistance;
   @override
-  double computeDistanceToActualBaseline(TextBaseline baseline) {
+  double? computeDistanceToActualBaseline(TextBaseline baseline) {
     // returns the baseline of the first cell that has a baseline in the first row
     assert(!debugNeedsLayout);
     return _baselineDistance;
@@ -791,7 +796,7 @@ class RenderTable extends RenderBox {
   Iterable<RenderBox> column(int x) sync* {
     for (int y = 0; y < rows; y += 1) {
       final int xy = x + y * columns;
-      final RenderBox child = _children[xy];
+      final RenderBox? child = _children[xy];
       if (child != null)
         yield child;
     }
@@ -805,7 +810,7 @@ class RenderTable extends RenderBox {
     final int start = y * columns;
     final int end = (y + 1) * columns;
     for (int xy = start; xy < end; xy += 1) {
-      final RenderBox child = _children[xy];
+      final RenderBox? child = _children[xy];
       if (child != null)
         yield child;
     }
@@ -825,9 +830,9 @@ class RenderTable extends RenderBox {
     //    necessary, applying minimum column widths as we go
 
     // 1. apply ideal widths, and collect information we'll need later
-    final List<double> widths = List<double>(columns);
-    final List<double> minWidths = List<double>(columns);
-    final List<double> flexes = List<double>(columns);
+    final List<double> widths = List<double>.filled(columns, 0.0, growable: false);
+    final List<double> minWidths = List<double>.filled(columns, 0.0, growable: false);
+    final List<double?> flexes = List<double?>.filled(columns, null, growable: false);
     double tableWidth = 0.0; // running tally of the sum of widths[x] for all x
     double unflexedTableWidth = 0.0; // sum of the maxIntrinsicWidths of any column that has null flex
     double totalFlex = 0.0;
@@ -847,17 +852,16 @@ class RenderTable extends RenderBox {
       minWidths[x] = minIntrinsicWidth;
       assert(maxIntrinsicWidth >= minIntrinsicWidth);
       // collect flex information while we're at it
-      final double flex = columnWidth.flex(columnCells);
+      final double? flex = columnWidth.flex(columnCells);
       if (flex != null) {
         assert(flex.isFinite);
         assert(flex > 0.0);
         flexes[x] = flex;
         totalFlex += flex;
       } else {
-        unflexedTableWidth += maxIntrinsicWidth;
+        unflexedTableWidth = unflexedTableWidth + maxIntrinsicWidth;
       }
     }
-    assert(!widths.any((double value) => value == null));
     final double maxWidthConstraint = constraints.maxWidth;
     final double minWidthConstraint = constraints.minWidth;
 
@@ -878,7 +882,7 @@ class RenderTable extends RenderBox {
         assert(remainingWidth >= 0.0);
         for (int x = 0; x < columns; x += 1) {
           if (flexes[x] != null) {
-            final double flexedWidth = remainingWidth * flexes[x] / totalFlex;
+            final double flexedWidth = remainingWidth * flexes[x]! / totalFlex;
             assert(flexedWidth.isFinite);
             assert(flexedWidth >= 0.0);
             if (widths[x] < flexedWidth) {
@@ -897,15 +901,11 @@ class RenderTable extends RenderBox {
     else if (tableWidth < minWidthConstraint) {
       final double delta = (minWidthConstraint - tableWidth) / columns;
       for (int x = 0; x < columns; x += 1)
-        widths[x] += delta;
+        widths[x] = widths[x] + delta;
       tableWidth = minWidthConstraint;
     }
 
     // beyond this point, unflexedTableWidth is no longer valid
-    assert(() {
-      unflexedTableWidth = null;
-      return true;
-    }());
 
     // 4. apply the maximum width of the table, shrinking columns as
     //    necessary, applying minimum column widths as we go
@@ -932,7 +932,7 @@ class RenderTable extends RenderBox {
         double newTotalFlex = 0.0;
         for (int x = 0; x < columns; x += 1) {
           if (flexes[x] != null) {
-            final double newWidth = widths[x] - deficit * flexes[x] / totalFlex;
+            final double newWidth = widths[x] - deficit * flexes[x]! / totalFlex;
             assert(newWidth.isFinite);
             if (newWidth <= minWidths[x]) {
               // shrank to minimum
@@ -943,7 +943,7 @@ class RenderTable extends RenderBox {
             } else {
               deficit -= widths[x] - newWidth;
               widths[x] = newWidth;
-              newTotalFlex += flexes[x];
+              newTotalFlex += flexes[x]!;
             }
             assert(widths[x] >= 0.0);
           }
@@ -968,7 +968,7 @@ class RenderTable extends RenderBox {
               widths[x] = minWidths[x];
             } else {
               deficit -= delta;
-              widths[x] -= delta;
+              widths[x] = widths[x] - delta;
               newAvailableColumns += 1;
             }
           }
@@ -981,7 +981,7 @@ class RenderTable extends RenderBox {
 
   // cache the table geometry for painting purposes
   final List<double> _rowTops = <double>[];
-  Iterable<double> _columnLefts;
+  Iterable<double>? _columnLefts;
 
   /// Returns the position and dimensions of the box that the given
   /// row covers, in this render object's coordinate space (so the
@@ -1010,7 +1010,7 @@ class RenderTable extends RenderBox {
       return;
     }
     final List<double> widths = _computeColumnWidths(constraints);
-    final List<double> positions = List<double>(columns);
+    final List<double> positions = List<double>.filled(columns, 0.0, growable: false);
     double tableWidth;
     switch (textDirection) {
       case TextDirection.rtl:
@@ -1028,7 +1028,6 @@ class RenderTable extends RenderBox {
         tableWidth = positions.last + widths.last;
         break;
     }
-    assert(!positions.any((double value) => value == null));
     _rowTops.clear();
     _baselineDistance = null;
     // then, lay out each row
@@ -1039,10 +1038,10 @@ class RenderTable extends RenderBox {
       bool haveBaseline = false;
       double beforeBaselineDistance = 0.0;
       double afterBaselineDistance = 0.0;
-      final List<double> baselines = List<double>(columns);
+      final List<double> baselines = List<double>.filled(columns, 0.0, growable: false);
       for (int x = 0; x < columns; x += 1) {
         final int xy = x + y * columns;
-        final RenderBox child = _children[xy];
+        final RenderBox? child = _children[xy];
         if (child != null) {
           final TableCellParentData childParentData = child.parentData as TableCellParentData;
           assert(childParentData != null);
@@ -1052,7 +1051,7 @@ class RenderTable extends RenderBox {
             case TableCellVerticalAlignment.baseline:
               assert(textBaseline != null);
               child.layout(BoxConstraints.tightFor(width: widths[x]), parentUsesSize: true);
-              final double childBaseline = child.getDistanceToBaseline(textBaseline, onlyReal: true);
+              final double? childBaseline = child.getDistanceToBaseline(textBaseline!, onlyReal: true);
               if (childBaseline != null) {
                 beforeBaselineDistance = math.max(beforeBaselineDistance, childBaseline);
                 afterBaselineDistance = math.max(afterBaselineDistance, child.size.height - childBaseline);
@@ -1081,7 +1080,7 @@ class RenderTable extends RenderBox {
       }
       for (int x = 0; x < columns; x += 1) {
         final int xy = x + y * columns;
-        final RenderBox child = _children[xy];
+        final RenderBox? child = _children[xy];
         if (child != null) {
           final TableCellParentData childParentData = child.parentData as TableCellParentData;
           switch (childParentData.verticalAlignment ?? defaultVerticalAlignment) {
@@ -1113,18 +1112,18 @@ class RenderTable extends RenderBox {
   }
 
   @override
-  bool hitTestChildren(BoxHitTestResult result, { Offset position }) {
+  bool hitTestChildren(BoxHitTestResult result, { required Offset position }) {
     assert(_children.length == rows * columns);
     for (int index = _children.length - 1; index >= 0; index -= 1) {
-      final RenderBox child = _children[index];
+      final RenderBox? child = _children[index];
       if (child != null) {
         final BoxParentData childParentData = child.parentData as BoxParentData;
         final bool isHit = result.addWithPaintOffset(
           offset: childParentData.offset,
           position: position,
-          hitTest: (BoxHitTestResult result, Offset transformed) {
+          hitTest: (BoxHitTestResult result, Offset? transformed) {
             assert(transformed == position - childParentData.offset);
-            return child.hitTest(result, position: transformed);
+            return child.hitTest(result, position: transformed!);
           },
         );
         if (isHit)
@@ -1140,20 +1139,20 @@ class RenderTable extends RenderBox {
     if (rows * columns == 0) {
       if (border != null) {
         final Rect borderRect = Rect.fromLTWH(offset.dx, offset.dy, size.width, 0.0);
-        border.paint(context.canvas, borderRect, rows: const <double>[], columns: const <double>[]);
+        border!.paint(context.canvas, borderRect, rows: const <double>[], columns: const <double>[]);
       }
       return;
     }
     assert(_rowTops.length == rows + 1);
     if (_rowDecorations != null) {
-      assert(_rowDecorations.length == _rowDecorationPainters.length);
+      assert(_rowDecorations!.length == _rowDecorationPainters!.length);
       final Canvas canvas = context.canvas;
       for (int y = 0; y < rows; y += 1) {
-        if (_rowDecorations.length <= y)
+        if (_rowDecorations!.length <= y)
           break;
-        if (_rowDecorations[y] != null) {
-          _rowDecorationPainters[y] ??= _rowDecorations[y].createBoxPainter(markNeedsPaint);
-          _rowDecorationPainters[y].paint(
+        if (_rowDecorations![y] != null) {
+          _rowDecorationPainters![y] ??= _rowDecorations![y]!.createBoxPainter(markNeedsPaint);
+          _rowDecorationPainters![y]!.paint(
             canvas,
             Offset(offset.dx, offset.dy + _rowTops[y]),
             configuration.copyWith(size: Size(size.width, _rowTops[y+1] - _rowTops[y])),
@@ -1162,22 +1161,22 @@ class RenderTable extends RenderBox {
       }
     }
     for (int index = 0; index < _children.length; index += 1) {
-      final RenderBox child = _children[index];
+      final RenderBox? child = _children[index];
       if (child != null) {
         final BoxParentData childParentData = child.parentData as BoxParentData;
         context.paintChild(child, childParentData.offset + offset);
       }
     }
     assert(_rows == _rowTops.length - 1);
-    assert(_columns == _columnLefts.length);
+    assert(_columns == _columnLefts!.length);
     if (border != null) {
       // The border rect might not fill the entire height of this render object
       // if the rows underflow. We always force the columns to fill the width of
       // the render object, which means the columns cannot underflow.
       final Rect borderRect = Rect.fromLTWH(offset.dx, offset.dy, size.width, _rowTops.last);
       final Iterable<double> rows = _rowTops.getRange(1, _rowTops.length - 1);
-      final Iterable<double> columns = _columnLefts.skip(1);
-      border.paint(context.canvas, borderRect, rows: rows, columns: columns);
+      final Iterable<double> columns = _columnLefts!.skip(1);
+      border!.paint(context.canvas, borderRect, rows: rows, columns: columns);
     }
   }
 
@@ -1189,7 +1188,7 @@ class RenderTable extends RenderBox {
     properties.add(DiagnosticsProperty<TableColumnWidth>('default column width', defaultColumnWidth));
     properties.add(MessageProperty('table size', '$columns\u00D7$rows'));
     properties.add(IterableProperty<String>('column offsets', _columnLefts?.map(debugFormatDouble), ifNull: 'unknown'));
-    properties.add(IterableProperty<String>('row offsets', _rowTops?.map(debugFormatDouble), ifNull: 'unknown'));
+    properties.add(IterableProperty<String>('row offsets', _rowTops.map(debugFormatDouble), ifNull: 'unknown'));
   }
 
   @override
@@ -1202,7 +1201,7 @@ class RenderTable extends RenderBox {
     for (int y = 0; y < rows; y += 1) {
       for (int x = 0; x < columns; x += 1) {
         final int xy = x + y * columns;
-        final RenderBox child = _children[xy];
+        final RenderBox? child = _children[xy];
         final String name = 'child ($x, $y)';
         if (child != null)
           children.add(child.toDiagnosticsNode(name: name));
