@@ -2,8 +2,6 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-// @dart = 2.8
-
 import 'dart:ui' as ui show Paragraph, ParagraphBuilder, ParagraphConstraints, ParagraphStyle, TextStyle;
 
 import 'box.dart';
@@ -59,7 +57,7 @@ class RenderErrorBox extends RenderBox {
   final String message;
 
   // TODO(ianh): should be final
-  /*late*/ ui.Paragraph/*?*/ _paragraph;
+  ui.Paragraph? _paragraph;
 
   @override
   double computeMaxIntrinsicWidth(double height) {
@@ -158,11 +156,11 @@ class RenderErrorBox extends RenderBox {
           width -= padding.left + padding.right;
           left += padding.left;
         }
-        _paragraph.layout(ui.ParagraphConstraints(width: width));
-        if (size.height > padding.top + _paragraph.height + padding.bottom) {
+        _paragraph!.layout(ui.ParagraphConstraints(width: width));
+        if (size.height > padding.top + _paragraph!.height + padding.bottom) {
           top += padding.top;
         }
-        context.canvas.drawParagraph(_paragraph, offset + Offset(left, top));
+        context.canvas.drawParagraph(_paragraph!, offset + Offset(left, top));
       }
     } catch (e) {
       // Intentionally left empty.
