@@ -253,9 +253,15 @@ class _GlowingOverscrollIndicatorState extends State<GlowingOverscrollIndicator>
     // scroll, e.g. when scrolling in the opposite direction again to hide
     // the glow. Otherwise, the glow would always stay in a fixed position,
     // even if the top of the content already scrolled away.
-    _leadingController._paintOffsetScrollPixels = -notification.metrics.pixels;
+    // For example (CustomScrollView with sliver before center), the scroll
+    // extent is [-200.0, 300.0], scroll in the opposite direction with 10.0 pixels
+    // before glow disappears, so the current pixels is -190.0,
+    // in this case, we should move the glow up 10.0 pixels and should not
+    // overflow the scrollable widget's edge. https://github.com/flutter/flutter/issues/64149.
+    _leadingController._paintOffsetScrollPixels =
+      -math.min(notification.metrics.pixels - notification.metrics.minScrollExtent, _leadingController._paintOffset);
     _trailingController._paintOffsetScrollPixels =
-        -(notification.metrics.maxScrollExtent - notification.metrics.pixels);
+      -math.min(notification.metrics.maxScrollExtent - notification.metrics.pixels, _trailingController._paintOffset);
 
     if (notification is OverscrollNotification) {
       _GlowController controller;
@@ -269,9 +275,6 @@ class _GlowingOverscrollIndicatorState extends State<GlowingOverscrollIndicator>
       final bool isLeading = controller == _leadingController;
       if (_lastNotificationType != OverscrollNotification) {
         final OverscrollIndicatorNotification confirmationNotification = OverscrollIndicatorNotification(leading: isLeading);
-        // It is possible that the scroll extent starts at non-zero.
-        if (isLeading)
-          confirmationNotification.paintOffset = notification.metrics.minScrollExtent;
         confirmationNotification.dispatch(context);
         _accepted[isLeading] = confirmationNotification._accepted;
         if (_accepted[isLeading]) {
